feat: mentoring program, portal settings hub, super-admin updates
Mentoring: milestones workspace, mentor/mentee cards, shared mentoring card preview
in Community settings and on mentee portal; RPC for mentor profile when embed is null;
import templates and admin mentoring tools; program-order rule for milestones.

Portal: settings shell and sections (account, pilot, community, commute, family view,
pay projection, subscription, etc.); profile/community forms; mentor contact fields.

Super Admin: user KPIs and live strip, online peak RPC, account deletion and pending
deletions flows, mentoring CSV and backfill controls.

Database: migrations through get_mentor_profile_for_mentee_card and online peak.

Misc: phone formatting helper, tenant airline display, soft-delete columns support,
Stripe and middleware touchpoints, package updates.

Made-with: Cursor

+++ app/frontier/pilots/portal/schedule/actions.ts
@@ -968,7 +968,7 @@ export async function getAvailableMonths(): Promise<MonthOption[]> {
 
     const toOption = (p: { index: number; name: string; startStr: string; endStr: string }): MonthOption => {
       const fmt = (s: string) =>
-        new Date(s + "T12:00:00.000Z").toLocaleString("en-US", { month: "short", day: "numeric" });
+        new Date(s + "T12:00:00.000Z").toLocaleString("en-US", { month: "long", day: "numeric" });
       return {
         year: currentYear,
         month: p.index,
